ported to 1.20.1 many things have been commented out - got it up and running first before fixing. many TODOs

+++ src/main/java/net/mokai/quicksandrehydrated/util/EasingHandler.java
@@ -51,7 +51,10 @@ public class EasingHandler {
 
         do {
             currentHeight++;
-            BlockPos check = new BlockPos(pPos.getX(), currentHeight, pPos.getZ());
+
+
+
+            BlockPos check = new BlockPos(pPos.getX(), (int) currentHeight, pPos.getZ());
             playercube = check;
         } while (pLevel.getBlockState(playercube).getBlock() instanceof QuicksandBase);
